Added security for interaction with server

+++ backend/events.js
@@ -26,9 +26,11 @@ module.exports = function (app, database, authenticate) {
 
         authenticate(request.get('Token'))
             .then((user) => {
+
                 if (user !== -1) {
                     database.all('SELECT * FROM users WHERE userId=?', [user])
                         .then((users) => {
+
                             database.run('INSERT INTO events (eventSport, eventTitle, eventDescription, eventGoers, eventUserId, eventUserName) VALUES (?, ?, ?, ?, ?, ?)',
                                 [
                                     request.body.eventSport,
@@ -58,16 +60,20 @@ module.exports = function (app, database, authenticate) {
                     database.all('SELECT * FROM events WHERE eventId=?', [request.params.event])
                         .then((events) => {
 
+                            if (user !== events[0].eventUserId) {
+                                response.send(JSON.stringify({ message: 'Unauthorized', status: 2 }))
+                                return
+                            }
+
                             events[0].eventGoers = JSON.parse(events[0].eventGoers)
 
                             let updatedEvent = Object.assign(events[0], request.body)
 
-                            database.run('UPDATE events SET eventSport=?, eventTitle=?, eventDescription=?, eventGoers=? WHERE eventId=?',
+                            database.run('UPDATE events SET eventSport=?, eventTitle=?, eventDescription=? WHERE eventId=?',
                                 [
                                     updatedEvent.eventSport,
                                     updatedEvent.eventTitle,
                                     updatedEvent.eventDescription,
-                                    JSON.stringify(updatedEvent.eventGoers),
                                     request.params.event
                                 ]).then(() => {
                                     response.send(JSON.stringify({ message: 'Event updated', status: 1 }))
@@ -81,34 +87,57 @@ module.exports = function (app, database, authenticate) {
 
     app.patch('/attend/:event', (request, response) => {
 
-        database.all('SELECT * FROM events WHERE eventId=?', [request.params.event])
-                        .then((events) => {
+        authenticate(request.get('Token'))
+            .then((user) => {
 
-                            events[0].eventGoers = JSON.parse(events[0].eventGoers)
+                if (user !== -1) {
+                    database.all('SELECT * FROM users WHERE userId=?', [user])
+                        .then((users) => {
 
-                            let updatedEvent = Object.assign(events[0], request.body)
+                            database.all('SELECT * FROM events WHERE eventId=?', [request.params.event])
+                                .then((events) => {
 
-                            database.run('UPDATE events SET eventSport=?, eventTitle=?, eventDescription=?, eventGoers=? WHERE eventId=?',
-                                [
-                                    updatedEvent.eventSport,
-                                    updatedEvent.eventTitle,
-                                    updatedEvent.eventDescription,
-                                    JSON.stringify(updatedEvent.eventGoers),
-                                    request.params.event
-                                ]).then(() => {
-                                    response.send(JSON.stringify({ message: 'Attending event', status: 1 }))
+                                    events[0].eventGoers = JSON.parse(events[0].eventGoers)
+
+                                    if (!request.body.attendeeExist) {
+                                        events[0].eventGoers.push(users[0].userName)
+                                    } else {
+                                        events[0].eventGoers.splice(events[0].eventGoers.indexOf(users[0].userName), 1)
+                                    }
+
+                                    database.run('UPDATE events SET eventGoers=? WHERE eventId=?',
+                                        [
+                                            JSON.stringify(events[0].eventGoers),
+                                            request.params.event
+                                        ]).then(() => {
+                                            response.send(JSON.stringify({ message: 'Attending event', status: 1 }))
+                                        })
                                 })
                         })
+                } else {
+                    response.send(JSON.stringify({ message: 'Unauthorized', status: 2 }))
+                }
+            })
     })
 
     app.delete('/events/:event', (request, response) => {
 
         authenticate(request.get('Token'))
             .then((user) => {
+
                 if (user !== -1) {
-                    database.run('DELETE FROM events WHERE eventId=?', [request.params.event])
-                        .then(() => {
-                            response.send(JSON.stringify({ message: 'Event deleted', status: 1 }))
+                    database.all('SELECT * FROM events WHERE eventId=?', [request.params.event])
+                        .then((events) => {
+
+                            if (user !== events[0].eventUserId) {
+                                response.send(JSON.stringify({ message: 'Unauthorized', status: 2 }))
+                                return
+                            }
+
+                            database.run('DELETE FROM events WHERE eventId=?', [request.params.event])
+                                .then(() => {
+                                    response.send(JSON.stringify({ message: 'Event deleted', status: 1 }))
+                                })
                         })
                 } else {
                     response.send(JSON.stringify({ message: 'Unauthorized', status: 2 }))
